# rust: Use u128s to copy blocks