This results in a noticible speedup.

Before:

    test bench::bench_aes_siv_128_encrypt_1024_bytes  ... bench:
    4,149 ns/iter (+/- 385) = 246 MB/s
    test bench::bench_aes_siv_128_encrypt_128_bytes   ... bench:
    649 ns/iter (+/- 78) = 197 MB/s
    test bench::bench_aes_siv_128_encrypt_16384_bytes ... bench:
    63,147 ns/iter (+/- 3,339) = 259 MB/s

After:

    test bench::bench_aes_siv_128_encrypt_1024_bytes  ... bench:
    3,279 ns/iter (+/- 216) = 312 MB/s
    test bench::bench_aes_siv_128_encrypt_128_bytes   ... bench:
    551 ns/iter (+/- 82) = 232 MB/s
    test bench::bench_aes_siv_128_encrypt_16384_bytes ... bench:
    51,337 ns/iter (+/- 6,315) = 319 MB/s